requirements.txt
django==6.0.1
django-unfold==0.77.1
pillow==12.1.0